requirements.txt
# 基本ライブラリ
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pydantic>=2.0.0

# スケジューリング
APScheduler>=3.10.0

# 通知
slack-sdk>=3.21.0
emails>=0.6.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# BigQuery（オプション）
google-cloud-bigquery>=3.11.0

# ログ・ユーティリティ
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0

# 開発用
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 